refactor: refactored activity architecture
