+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { UserScoreModule } from '../modules/user-score/user-score.module';
 import { CourseModule } from '../modules/course/infra/modules/course.module';
 import { UserActivitiesAnsweredModule } from '../modules/user-activities-answered/user-activities-answered.module';
 import { UserCoursesConcludedModule } from '../modules/user-courses-concluded/user-courses-concluded.module';
-import { ActivityModule } from '../modules/activity/activity.module';
+import { ActivityModule } from '../modules/activity/infra/modules/activity.module';
 
 @Module({
   imports: [
