+++ src/modules/activity/domain/dtos/errors/ActivityNotFound.exception.ts
