+++ src/modules/activity/domain/dtos/requests/FindActivitiesCollection.request.dto.ts
