--- src/modules/activity/dto/activity.dto.spec.ts
@@ -1,7 +1,0 @@
-import { ActivityDto } from './activity.dto';
-
-describe('ActivityDto', () => {
-  it('should be defined', () => {
-    expect(new ActivityDto()).toBeDefined();
-  });
-});
