--- src/modules/activity/dto/activity.dto.ts
@@ -1,1 +1,0 @@
-export class ActivityDto {}
