+++ src/modules/activity/infra/db/entity/activity.entity.ts
@@ -1,4 +1,4 @@
-import { Course } from '../../course/infra/db/entity/course.entity';
+import { Course } from '../../../../course/infra/db/entity/course.entity';
 import { User } from '../../../modules/user/entity/user.entity';
 import {
   Column,
