+++ src/modules/course/infra/modules/course.module.ts
@@ -11,11 +11,11 @@ import {
 import { CourseService } from '../services/course.service';
 import { DatabaseModule } from '../../../../database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Activity } from '../../../activity/entity/activity.entity';
+import { Activity } from '../../../activity/infra/db/entity/activity.entity';
 import { User } from '../../../user/infra/db/entity/user.entity';
 import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repository/user-courses-concluded.repository';
 import { CourseRepository } from '../db/repositories/course.repository';
-import { ActivityRepository } from '../../../activity/repository/activity.repository';
+import { ActivityRepository } from '../../../activity/infra/db/repositories/activity.repository';
 import { AuthenticationMiddleware } from '../../../user/infra/http/middlewares/auth.middleware';
 import { UserCourseConcluded } from '../../../user-courses-concluded/entity/user-courses-concluded.entity';
 import { JWTProvider } from '../../../user/infra/providers/jwt.provider';
