+++ src/modules/course/infra/services/course.service.spec.ts
@@ -2,13 +2,13 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { CourseService } from './course.service';
 import { DatabaseModule } from '../../../../database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../../user/entity/user.entity';
+import { User } from '../../../user/infra/db/entity/user.entity';
 import { UserCourseConcluded } from '../../../user-courses-concluded/entity/user-courses-concluded.entity';
-import { Activity } from '../../../activity/entity/activity.entity';
-import { CourseRepository } from '../repositories/course.repository';
+import { Activity } from '../../../activity/infra/db/entity/activity.entity';
+import { CourseRepository } from '../db/repositories/course.repository';
 import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repository/user-courses-concluded.repository';
-import { ActivityRepository } from '../../../activity/repository/activity.repository';
-import { JWTProvider } from '../../user/providers/jwt.provider';
+import { ActivityRepository } from '../../../activity/infra/db/repositories/activity.repository';
+import { JWTProvider } from '../../../user/infra/providers/jwt.provider';
 import { UserActivityAnsweredRepository } from '../../../user-activities-answered/repository/user-activities-answered.repository';
 import {
   FindCoursesResponseDTO,
@@ -17,7 +17,7 @@ import {
 import { UserNotFoundException } from '../../../user/domain/dtos/errors/UserNotFound.exception';
 import { CourseNotFoundException } from '../../domain/dtos/errors/CourseNotFound.exception';
 import { UnprocessableDataException } from '../../../../shared/domain/errors/UnprocessableData.exception';
-import { UserRepository } from '../../user/repositories/user.repository';
+import { UserRepository } from '../../../user/infra/db/repositories/user.repository';
 import { TestHelper } from '../../../../../test/helpers/dbInstanceHelper';
 
 describe('Course Service Automated Tests', () => {
