+++ src/modules/user/infra/modules/user.module.ts
@@ -10,7 +10,7 @@ import { HashProvider } from '../providers/hash.provider';
 import { UserController } from '../http/controllers/user.controller';
 import { DatabaseModule } from '../../../../database/database.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Activity } from '../../../activity/entity/activity.entity';
+import { Activity } from '../../../activity/infra/db/entity/activity.entity';
 import { Course } from '../../../course/infra/db/entity/course.entity';
 import { UserCourseConcluded } from '../../../user-courses-concluded/entity/user-courses-concluded.entity';
 import { UserActivityAnswered } from '../../../user-activities-answered/entity/user-activities-answered.entity';
